Time: 182 ms (86.53%), Space: 79.7 MB (50.62%) - LeetHub

+++ 2616-maximal-score-after-applying-k-operations/2616-maximal-score-after-applying-k-operations.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long maxKelements(vector<int>& nums, int k) {
-        priority_queue<long long> pq;
+        priority_queue<int> pq;
         for(int i=0 ; i<nums.size() ; i++){
             pq.push(nums[i]);
         }
